+++ publish.sh
@@ -2,8 +2,8 @@
 
 set -e # exit with nonzero exit code if anything fails
 
-openssl aes-256-cbc -K $encrypted_6c42794c35b3_key -iv $encrypted_6c42794c35b3_iv -in secrets.tar.enc -out secrets.tar -d
+openssl aes-256-cbc -K $encrypted_key -iv $encrypted_iv -in secrets.tar.enc -out secrets.tar -d
 
 tar xvf secrets.tar
 
-sbt publishSignedAll sonatypeRelease
+sbt publishSignedAll coreIntellijSupport/updateIntellij coreIntellijSupport/publishSigned sonatypeBundleRelease