+++ CHANGELOG.md
@@ -2,7 +2,18 @@
 
 ## vNEXT
 
-## [v0.6.2](https://slinky.dev)
+## [v0.6.3](https://slinky.dev)
+### Highlights :tada:
++ Improve experience using `ReactElement`s within first-order types such as Map and List [PR #285](https://github.com/shadaj/slinky/pull/285)
++ Support [React Native's Keyboard API](https://facebook.github.io/react-native/docs/keyboard) [PR #293](https://github.com/shadaj/slinky/pull/293)
+
+### Bug Fixes
++ Fix crashes when calling `setState` on a component with `Props` set to some subtype of `Function1` [PR #295](https://github.com/shadaj/slinky/pull/295)
++ Support `useCallback` with a function that takes arguments [PR #290](https://github.com/shadaj/slinky/pull/290)
++ Fix false warnings on `@react` classes if `-Ywarn-value-discard` scalac option was used [PR #296](https://github.com/shadaj/slinky/pull/296)
++ Update the IntelliJ support plugin to be compatible with IntelliJ 2019.2 [PR #297](https://github.com/shadaj/slinky/pull/297)
+
+## [v0.6.2](https://5d1e6bd54bf73a00097195a2--slinky.netlify.com/)
 ### Highlights :tada:
 + Support **optional attributes**, see [the docs](https://slinky.dev/docs/the-tag-api/) for more details [PR #275](https://github.com/shadaj/slinky/pull/275)
 + Slinky now has **full support for Scala 2.13** [PR #269](https://github.com/shadaj/slinky/pull/269)